Move updateSnapshots to OnyxUtils
